Corrección relación cliente y usuario, y sus métodos :)

+++ src/main/java/ec/edu/ups/BoscoMarketApi/entidades/peticiones/Usuario/CrearUsuario.java
@@ -16,7 +16,7 @@ public class CrearUsuario {
     private String rol;
     @JsonProperty
     @Getter @Setter
-    private int cliente;
+    private long cliente;
 
     public CrearUsuario() {
     }
