+++ src/main/java/ec/edu/ups/BoscoMarketApi/repositorios/ClienteRepositorio.java
@@ -3,5 +3,5 @@ package ec.edu.ups.BoscoMarketApi.repositorios;
 import ec.edu.ups.BoscoMarketApi.entidades.Cliente;
 import org.springframework.data.repository.CrudRepository;
 
-public interface ClienteRepositorio extends CrudRepository<Cliente, Integer> {
+public interface ClienteRepositorio extends CrudRepository<Cliente, Long> {
 }
